fix: use environment variables for sensitive data
